Added Efficient Download mechanism to reuse already downloaded package

+++ vcs_win.h
@@ -213,6 +213,33 @@ void check_codeblocks()
     codeblocks = isFileExists("C:\\Program Files\\CodeBlocks\\codeblocks.exe");
 }
 
+void press_Y_to_continue_exe()
+{
+    char ch;
+    while (1)
+    {
+        printf("Proceed with installation....\n");
+        printf("If complete enter 'Y' to continue....\n");
+        ch = getchar();
+        if (ch == 'Y' || ch == 'y')
+            break;
+        system("cls");
+    }
+}
+void press_Y_to_continue_script()
+{
+    char ch;
+    while (1)
+    {
+        printf("Proceed with installation....\n");
+        printf("If complete enter 'Y' to continue....\n");
+        ch = getchar();
+        if (ch == 'Y' || ch == 'y')
+            break;
+        system("cls");
+    }
+}
+
 void construct_download_list()
 {
     FILE *fp;
@@ -246,9 +273,12 @@ int print_system_info()
 void Download_file_wget()
 {
     // system(".\\wget.exe \"https://sourceforge.net/projects/mingw-w64/files/Toolchains%20targetting%20Win64/Personal%20Builds/mingw-builds/8.1.0/threads-posix/seh/x86_64-8.1.0-release-posix-seh-rt_v6-rev0.7z\" -O gcc-latest.7z");
-    system(".\\wget.exe \"https://github.com/msys2/msys2-installer/releases/download/2021-11-30/msys2-x86_64-20211130.exe \" -O msys2-latest.exe");
-    system(".\\wget.exe \"https://code.visualstudio.com/sha/download?build=stable&os=win32-x64-user\" -O vscode-latest.exe");
-    system(".\\wget.exe \"https://sourceforge.net/projects/codeblocks/files/Binaries/20.03/Windows/codeblocks-20.03mingw-setup.exe/download\" -O codeblocks-latest.exe");
+    if (!isFileExists(".\\msys2-latest.exe\""))
+        system(".\\wget.exe \"https://github.com/msys2/msys2-installer/releases/download/2021-11-30/msys2-x86_64-20211130.exe \" -O msys2-latest.exe");
+    if (!isFileExists(".\\vscode-latest.exe"))
+        system(".\\wget.exe \"https://code.visualstudio.com/sha/download?build=stable&os=win32-x64-user\" -O vscode-latest.exe");
+    if (!isFileExists(".\\codeblocks-latest.exe"))
+        system(".\\wget.exe \"https://sourceforge.net/projects/codeblocks/files/Binaries/20.03/Windows/codeblocks-20.03mingw-setup.exe/download\" -O codeblocks-latest.exe");
 }
 
 void app_exec()
